:recycle: #15 ReactのPageコンポーネントに関係ないCognitoのToken検証処理を別の関数として切り出し
